revision 0.2.4: Don't deadlock on subprocess.call
subprocess.call() makes use of Popen.call() internally [1] which is considered
bad in the documentation [2] because it can result in a deadlock if too much
data is read. The sourcecode of subprocess.call() suggests that a simple call
to `Popen(cmd, shell=True).communicate()` would produce the same result as
subprocess.call() but without that bug [1]. Therefore this is the currently
favored solution.

[1] http://hg.python.org/cpython/file/374746c5dedc/Lib/subprocess.py#l522
[2] https://docs.python.org/2/library/subprocess.html#subprocess.Popen.wait

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>